выделил genre в отдельный компонент
